# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: 紫式部
    url: https://discord.gg/KEfVkfC6Q9
  description: |-
    Sonolusの基本APIを拡張する感じ。
    ユーザー認証はFirebaseAuthorizationを通してやる。
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  termsOfService: https://discord.gg/KEfVkfC6Q9
  title: SweetPotato Server API
  version: "1.0"
servers:
- description: Local
  url: http://localhost:5000
...
paths:
...
  /effects/list:
    get:
      description: |-
        It returns list of effect infos registered in this server
        Also it can search using query params
      operationId: getEffectList
      parameters:
      - description: It localizes response items if possible
        explode: true
        in: query
        name: localization
        required: false
        schema:
          example: en
          maxLength: 50
          minLength: 1
          type: string
        style: form
      - description: It filters items for pagination if possible
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          maximum: 10000
          minimum: 0
          type: number
        style: form
      - description: It filters items for search from list if possible
        explode: true
        in: query
        name: keywords
        required: false
        schema:
          example: Redo
          maxLength: 300
          minLength: 1
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEffectListResponse'
          description: OK
      summary: Get effect list
      tags:
      - effects
  /effects/{effectName}:
    get:
      description: |-
        It returns specified effect info
        It will raise 404 if the effect is not registered in this server
      operationId: getEffect
      parameters:
      - explode: false
        in: path
        name: effectName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEffectResponse'
          description: OK
        "404":
          description: Not Found
      summary: Get effect
      tags:
      - effects
    patch:
      description: 指定されたeffectを編集します
      operationId: editEffect
      parameters:
      - explode: false
        in: path
        name: effectName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Effect'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: Edit effect
      tags:
      - effects
    post:
      description: 指定されたeffectをサーバーに登録します
      operationId: addEffect
      parameters:
      - explode: false
        in: path
        name: effectName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Effect'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "409":
          description: Conflict
      summary: Add effect
      tags:
      - effects
...
  /particles/list:
    get:
      description: |-
        It returns list of particle infos registered in this server
        Also it can search using query params
      operationId: getParticleList
      parameters:
      - description: It localizes response items if possible
        explode: true
        in: query
        name: localization
        required: false
        schema:
          example: en
          maxLength: 50
          minLength: 1
          type: string
        style: form
      - description: It filters items for pagination if possible
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          maximum: 10000
          minimum: 0
          type: number
        style: form
      - description: It filters items for search from list if possible
        explode: true
        in: query
        name: keywords
        required: false
        schema:
          example: Redo
          maxLength: 300
          minLength: 1
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetParticleListResponse'
          description: OK
      summary: Get particle list
      tags:
      - particles
  /particles/{particleName}:
    get:
      description: |-
        It returns specified particle info
        It will raise 404 if the particle is not registered in this server
      operationId: getParticle
      parameters:
      - explode: false
        in: path
        name: particleName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetParticleResponse'
          description: OK
        "404":
          description: Not Found
      summary: Get particle
      tags:
      - particles
    patch:
      description: 指定したparticleを編集します
      operationId: patch-particles-particleName
      parameters:
      - explode: false
        in: path
        name: particleName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Particle'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: Edit particle
      tags:
      - particles
    post:
      description: 指定されたパーティクル情報をサーバーに登録します
      operationId: addParticle
      parameters:
      - explode: false
        in: path
        name: particleName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Particle'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "409":
          description: Conflict
      summary: Add particle
      tags:
      - particles
...
components:
...
  schemas:
...
    Effect:
      description: |-
        An effect provides audio elements to levels / It defines audio effect used for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect
      example:
        thumbnail:
          type: LevelData
          hash: hash
          url: url
        data:
          clips:
          - id: 1
            clip:
              type: LevelData
              hash: hash
              url: url
          - id: 1
            clip:
              type: LevelData
              hash: hash
              url: url
        author: author
        subtitle: subtitle
        name: name
        title: title
        version: 1
      properties:
        name:
          description: english and number only name for searching
          maxLength: 100
          minLength: 1
          type: string
        version:
          default: 1
          description: Reserved for future update. current default is 1.
          example: 1
          maximum: 100
          minimum: 1
          type: integer
        title:
          description: base title of this content
          maxLength: 100
          minLength: 1
          type: string
        subtitle:
          description: something footer(ex. featuring xyz) for this content
          maxLength: 50
          minLength: 1
          type: string
        author:
          description: author of this content
          minLength: 1
          type: string
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/EffectData'
      required:
      - author
      - data
      - name
      - subtitle
      - thumbnail
      - title
      - version
      title: Effect
      type: object
      x-examples: {}
    Particle:
      description: |-
        A particle provides particle effect elements to levels / It defines particle effect used for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Particle
      example:
        thumbnail:
          type: LevelData
          hash: hash
          url: url
        data:
          type: LevelData
          hash: hash
          url: url
        author: author
        texture:
          type: LevelData
          hash: hash
          url: url
        subtitle: subtitle
        name: name
        title: title
        version: 1
      properties:
        name:
          description: english and number only name for searching
          maxLength: 100
          minLength: 1
          type: string
        version:
          default: 1
          description: Reserved for future update. current default is 1.
          example: 1
          maximum: 100
          minimum: 1
          type: integer
        title:
          description: base title of this content
          maxLength: 100
          minLength: 1
          type: string
        subtitle:
          description: something footer(ex. featuring xyz) for this content
          maxLength: 100
          minLength: 1
          type: string
        author:
          description: author of this content
          maxLength: 50
          minLength: 1
          type: string
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        texture:
          $ref: '#/components/schemas/SonolusResourceLocator'
      required:
      - author
      - data
      - name
      - subtitle
      - texture
      - thumbnail
      - title
      - version
      title: Particle
      type: object
      x-examples: {}
...
    EffectData:
      description: |-
        Effect clips are audio elements of an effect / It defines effect types and audio file locations
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect#data
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect-Clips
      example:
        clips:
        - id: 1
          clip:
            type: LevelData
            hash: hash
            url: url
        - id: 1
          clip:
            type: LevelData
            hash: hash
            url: url
      properties:
        clips:
          items:
            $ref: '#/components/schemas/EffectData_clips'
          minItems: 1
          type: array
          uniqueItems: true
      required:
      - clips
      title: EffectData
      type: object
      x-examples:
        example-1:
          clips:
          - id: 1
            clip: https://example.com/se-audio.mp3
...
    SonolusResourceLocator:
      description: |-
        Sonolus Resource Locator(SRL) is used in server endpoints in order for Sonolus to locate a resource.
        All resources must be gzip compressed.
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Resource-and-SRL
      example:
        type: LevelData
        hash: hash
        url: url
      properties:
        type:
          description: When specify type together with hash uniquely identifies a
            resource.
          enum:
          - SkinData
          - EngineThumbnail
          - EffectData
          - SkinTexture
          - ParticleData
          - EffectThumbnail
          - LevelBgm
          - BackgroundThumbnail
          - EngineConfiguration
          - LevelData
          - ParticleTexture
          - BackgroundData
          - EngineData
          - BackgroundImage
          - LevelCover
          - SkinThumbnail
          - ParticleThumbnail
          example: LevelData
          maxLength: 50
          minLength: 1
          type: string
        hash:
          description: Hash is calculated using SHA1 on the resource file
          maxLength: 100
          minLength: 1
          type: string
        url:
          description: If url starts with character /, it will be treated as relative
            path to server address.
          maxLength: 500
          minLength: 1
          type: string
      required:
      - url
      title: SonolusResourceLocator
      type: object
      x-examples:
        example:
          type: EffectData
          hash: addb1eb6402aa0dcb0d65c5a8dfd3935f08e3d39
          url: /repository/EffectData/addb1eb6402aa0dcb0d65c5a8dfd3935f08e3d39
...
    GetEffectListResponse:
      description: Response struct of getEffectList
      example:
        pageCount: 1
        items:
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            clips:
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
          author: author
          subtitle: subtitle
          name: name
          title: title
          version: 1
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            clips:
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
          author: author
          subtitle: subtitle
          name: name
          title: title
          version: 1
      properties:
        pageCount:
          default: 1
          description: used for pagination, default is 1
          example: 1
          minimum: 1
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Effect'
          type: array
      required:
      - items
      - pageCount
      title: GetEffectListResponse
      type: object
      x-examples: {}
...
    GetParticleListResponse:
      description: Response struct of getParticleList
      example:
        pageCount: 1
        items:
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            type: LevelData
            hash: hash
            url: url
          author: author
          texture:
            type: LevelData
            hash: hash
            url: url
          subtitle: subtitle
          name: name
          title: title
          version: 1
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            type: LevelData
            hash: hash
            url: url
          author: author
          texture:
            type: LevelData
            hash: hash
            url: url
          subtitle: subtitle
          name: name
          title: title
          version: 1
      properties:
        pageCount:
          default: 1
          description: used for pagination, default is 1
          example: 1
          minimum: 1
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Particle'
          type: array
      required:
      - items
      - pageCount
      title: GetParticleListResponse
      type: object
      x-examples: {}
...
    GetEffectResponse:
      description: Response struct of getEffect
      example:
        item:
          thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            clips:
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
          author: author
          subtitle: subtitle
          name: name
          title: title
          version: 1
        description: description
        recommended:
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            clips:
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
          author: author
          subtitle: subtitle
          name: name
          title: title
          version: 1
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            clips:
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
            - id: 1
              clip:
                type: LevelData
                hash: hash
                url: url
          author: author
          subtitle: subtitle
          name: name
          title: title
          version: 1
      properties:
        item:
          $ref: '#/components/schemas/Effect'
        description:
          minLength: 1
          type: string
        recommended:
          items:
            $ref: '#/components/schemas/Effect'
          type: array
      required:
      - description
      - item
      - recommended
      title: GetEffectResponse
      type: object
      x-examples: {}
...
    GetParticleResponse:
      description: Response struct of getParticle
      example:
        item:
          thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            type: LevelData
            hash: hash
            url: url
          author: author
          texture:
            type: LevelData
            hash: hash
            url: url
          subtitle: subtitle
          name: name
          title: title
          version: 1
        description: description
        recommended:
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            type: LevelData
            hash: hash
            url: url
          author: author
          texture:
            type: LevelData
            hash: hash
            url: url
          subtitle: subtitle
          name: name
          title: title
          version: 1
        - thumbnail:
            type: LevelData
            hash: hash
            url: url
          data:
            type: LevelData
            hash: hash
            url: url
          author: author
          texture:
            type: LevelData
            hash: hash
            url: url
          subtitle: subtitle
          name: name
          title: title
          version: 1
      properties:
        item:
          $ref: '#/components/schemas/Particle'
        description:
          minLength: 1
          type: string
        recommended:
          items:
            $ref: '#/components/schemas/Particle'
          type: array
      required:
      - description
      - item
      - recommended
      title: GetParticleResponse
      type: object
      x-examples: {}
...
    EffectData_clips:
      description: Effect clips are audio elements of an effect.
      example:
        id: 1
        clip:
          type: LevelData
          hash: hash
          url: url
      properties:
        id:
          description: "These effect types are must be implemented.\n1\tnormal note\
            \ perfect hit sound\n2\tnormal note great hit sound\n3\tnormal note good\
            \ hit sound\n4\talternative note hit sound\n5\tstage hit sound"
          example: 1
          minimum: 1
          type: integer
        clip:
          $ref: '#/components/schemas/SonolusResourceLocator'
      required:
      - clip
      - id
      type: object
...
  securitySchemes: {}