# gen/api_effects.go
/*
 * SweetPotato Server API
 *
 * Sonolusの基本APIを拡張する感じ。 ユーザー認証はFirebaseAuthorizationを通してやる。
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gen

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A EffectsApiController binds http requests to an api service and writes the service results to the http response
type EffectsApiController struct {
	service EffectsApiServicer
}

// NewEffectsApiController creates a default api controller
func NewEffectsApiController(s EffectsApiServicer) Router {
	return &EffectsApiController{service: s}
}

// Routes returns all of the api route for the EffectsApiController
func (c *EffectsApiController) Routes() Routes {
	return Routes{ 
		{
			"AddEffect",
			strings.ToUpper("Post"),
			"/effects/{effectName}",
			c.AddEffect,
		},
		{
			"EditEffect",
			strings.ToUpper("Patch"),
			"/effects/{effectName}",
			c.EditEffect,
		},
		{
			"GetEffect",
			strings.ToUpper("Get"),
			"/effects/{effectName}",
			c.GetEffect,
		},
		{
			"GetEffectList",
			strings.ToUpper("Get"),
			"/effects/list",
			c.GetEffectList,
		},
	}
}

// AddEffect - Add effect
func (c *EffectsApiController) AddEffect(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	effectName := params["effectName"]
	
	effect := &Effect{}
	if err := json.NewDecoder(r.Body).Decode(&effect); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	result, err := c.service.AddEffect(r.Context(), effectName, *effect)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// EditEffect - Edit effect
func (c *EffectsApiController) EditEffect(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	effectName := params["effectName"]
	
	effect := &Effect{}
	if err := json.NewDecoder(r.Body).Decode(&effect); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	result, err := c.service.EditEffect(r.Context(), effectName, *effect)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetEffect - Get effect
func (c *EffectsApiController) GetEffect(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	effectName := params["effectName"]
	
	result, err := c.service.GetEffect(r.Context(), effectName)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetEffectList - Get effect list
func (c *EffectsApiController) GetEffectList(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	localization := query.Get("localization")
	page := query.Get("page")
	keywords := query.Get("keywords")
	result, err := c.service.GetEffectList(r.Context(), localization, page, keywords)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}