# gen/api_particles.go
/*
 * SweetPotato Server API
 *
 * Sonolusの基本APIを拡張する感じ。 ユーザー認証はFirebaseAuthorizationを通してやる。
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gen

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A ParticlesApiController binds http requests to an api service and writes the service results to the http response
type ParticlesApiController struct {
	service ParticlesApiServicer
}

// NewParticlesApiController creates a default api controller
func NewParticlesApiController(s ParticlesApiServicer) Router {
	return &ParticlesApiController{service: s}
}

// Routes returns all of the api route for the ParticlesApiController
func (c *ParticlesApiController) Routes() Routes {
	return Routes{ 
		{
			"AddParticle",
			strings.ToUpper("Post"),
			"/particles/{particleName}",
			c.AddParticle,
		},
		{
			"GetParticle",
			strings.ToUpper("Get"),
			"/particles/{particleName}",
			c.GetParticle,
		},
		{
			"GetParticleList",
			strings.ToUpper("Get"),
			"/particles/list",
			c.GetParticleList,
		},
		{
			"PatchParticlesParticleName",
			strings.ToUpper("Patch"),
			"/particles/{particleName}",
			c.PatchParticlesParticleName,
		},
	}
}

// AddParticle - Add particle
func (c *ParticlesApiController) AddParticle(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	particleName := params["particleName"]
	
	particle := &Particle{}
	if err := json.NewDecoder(r.Body).Decode(&particle); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	result, err := c.service.AddParticle(r.Context(), particleName, *particle)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetParticle - Get particle
func (c *ParticlesApiController) GetParticle(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	particleName := params["particleName"]
	
	result, err := c.service.GetParticle(r.Context(), particleName)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetParticleList - Get particle list
func (c *ParticlesApiController) GetParticleList(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	localization := query.Get("localization")
	page := query.Get("page")
	keywords := query.Get("keywords")
	result, err := c.service.GetParticleList(r.Context(), localization, page, keywords)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// PatchParticlesParticleName - Edit particle
func (c *ParticlesApiController) PatchParticlesParticleName(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	particleName := params["particleName"]
	
	particle := &Particle{}
	if err := json.NewDecoder(r.Body).Decode(&particle); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	result, err := c.service.PatchParticlesParticleName(r.Context(), particleName, *particle)
	// If an error occurred, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}