# gen/model_effect_data.go
/*
 * SweetPotato Server API
 *
 * Sonolusの基本APIを拡張する感じ。 ユーザー認証はFirebaseAuthorizationを通してやる。
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gen

// EffectData - Effect clips are audio elements of an effect / It defines effect types and audio file locations https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect#data https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect-Clips
type EffectData struct {

	Clips []EffectDataClips `json:"clips"`
}