# gen/model_get_particle_list_response.go
/*
 * SweetPotato Server API
 *
 * Sonolusの基本APIを拡張する感じ。 ユーザー認証はFirebaseAuthorizationを通してやる。
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gen

// GetParticleListResponse - Response struct of getParticleList
type GetParticleListResponse struct {

	// used for pagination, default is 1
	PageCount int32 `json:"pageCount"`

	Items []Particle `json:"items"`
}