# gen/model_sonolus_resource_locator.go
/*
 * SweetPotato Server API
 *
 * Sonolusの基本APIを拡張する感じ。 ユーザー認証はFirebaseAuthorizationを通してやる。
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gen

// SonolusResourceLocator - Sonolus Resource Locator(SRL) is used in server endpoints in order for Sonolus to locate a resource. All resources must be gzip compressed. https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Resource-and-SRL
type SonolusResourceLocator struct {

	// When specify type together with hash uniquely identifies a resource.
	Type string `json:"type,omitempty"`

	// Hash is calculated using SHA1 on the resource file
	Hash string `json:"hash,omitempty"`

	// If url starts with character /, it will be treated as relative path to server address.
	Url string `json:"url"`
}